# internal/app/api/customer.go
package api

import (
	"github.com/ak1m1tsu/jokerge/internal/pkg/types"
	"github.com/gofiber/fiber/v2"
	"github.com/rs/zerolog"
)

// CustomerList возвращает список клиентов
//
//	@Summary	список клиентов
//	@Tags		customers
//	@Produce	json
//	@Param		X-Request-ID	header		string	true	"ID запроса"
//	@Success	200				{object}	types.CustomerListResponse
//	@Failure	500				{object}	types.APIResponse
//	@Router		/api/v1/customer/list [get]
func (e *Env) CustomerList(ctx *fiber.Ctx) error {
	customers, err := e.Service().GetCustomers(ctx.UserContext())
	if err != nil {
		zerolog.Ctx(ctx.UserContext()).Error().Err(err).Msg("failed to get customer list")
		return err
	}

	response := make(types.CustomerListResponse, 0, len(customers))

	for _, c := range customers {
		converted := types.CustomerInfoResponse{
			ID:        c.ID,
			FirstName: c.FirstName,
			LastName:  c.LastName,
			Address:   c.Address,
			Orders:    make([]types.CustomerInfoOrderResponse, 0, len(c.Orders)),
		}

		for _, o := range c.Orders {
			converted.Orders = append(converted.Orders, types.CustomerInfoOrderResponse{
				ID:        o.ID,
				Status:    o.Status.String(),
				Price:     o.CalculatePrice(),
				CreatedAt: o.CreatedAt.String(),
			})
		}

		response = append(response, converted)

	}

	return ctx.JSON(response)
}

// CustomerGet возвращает информацию о клиенте
//
//	@Summary	информация о клиенте
//	@Tags		customers
//	@Produce	json
//	@Accept		json
//	@Param		customer_id		path		string	true	"ID клиента"
//	@Param		X-Request-ID	header		string	true	"ID запроса"
//	@Success	200				{object}	types.CustomerInfoResponse
//	@Failure	404				{object}	types.APIResponse
//	@Failure	500				{object}	types.APIResponse
//	@Router		/api/v1/customer/{customer_id} [get]
func (e *Env) CustomerGet(ctx *fiber.Ctx) error {
	id := ctx.Params("id")

	customer, err := e.Service().GetCustomerInfo(ctx.UserContext(), id)
	if err != nil {
		zerolog.Ctx(ctx.UserContext()).Error().Err(err).Msgf("failed to get customer by id %s", id)
		return err
	}

	if customer == nil {
		return e.NotFound(ctx)
	}

	return ctx.JSON(customer)
}

// CustomerCreate создает клиента
//
//	@Summary	создание нового клиента
//	@Tags		customers
//	@Accept		json
//	@Produce	json
//	@Param		X-Request-ID	header		string	true	"ID запроса"
//	@Success	200				{object}	types.APIResponse
//	@Failure	400				{object}	types.APIResponse
//	@Failure	500				{object}	types.APIResponse
//	@Router		/api/v1/customer [post]
func (e *Env) CustomerCreate(ctx *fiber.Ctx) error {
	return e.OK(ctx)
}